% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_funs.R
\name{check_proportion}
\alias{check_proportion}
\title{check_proportion}
\usage{
check_proportion(.results, .name = "name", .pct = "pct")
}
\arguments{
\item{.results}{Dataframe or tibble. Output from
\code{\link{tabulate_results}}.}

\item{.name}{String. The column name containing nominees' names.}

\item{.pct}{String. The column name containing nominees' first-choice vote
proportions.}
}
\value{
A numeric scalar. The maximum first-choice vote proportion among
nominees.
}
\description{
Check the maximum proportion of first-choice votes for remaining nominees.
}
