% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomination_funs.R
\name{count_nominees}
\alias{count_nominees}
\title{count_nominees}
\usage{
count_nominees(.data, .limit = NULL, col_name = "Nominees")
}
\arguments{
\item{.data}{Dataframe/tibble. The data output by
\code{\link{import_nominations}}.}

\item{.limit}{Either \code{NULL} (if no limit desired) or a numeric scalar
if output nominations should be limited to only the most frequently nominated
nominees. For example, if \code{.limit = 5}, then only the nominees with
the top 5 nomination counts will be output. This preserves ties in case
several nominees are nominated the same number of times.}

\item{col_name}{String. Name to be passed to the column of nominees.}
}
\value{
A tibble of nominee names and nomination counts.
}
\description{
count_nominees
}
