% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_funs.R
\name{detect_lowest}
\alias{detect_lowest}
\title{detect_lowest}
\usage{
detect_lowest(.results, .name = "name", .pct = "pct")
}
\arguments{
\item{.results}{Dataframe or tibble. Output from
\code{\link{tabulate_results}}.}

\item{.name}{String. The column name containing nominees' names.}

\item{.pct}{String. The column name containing nominees' first-choice vote
proportions.}
}
\value{
A string. The name of the nominee with the lowest number of
first-choice votes.
}
\description{
Find the nominee with the lowest number of first-choice votes (input is
output of \code{\link{tabulate_results}} function).
}
