% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_funs.R
\name{find_max_result}
\alias{find_max_result}
\title{print_max_results}
\usage{
find_max_result(.results, .majority, .name = "name", .pct = "pct")
}
\arguments{
\item{.results}{Dataframe or tibble. Output from
\code{\link{tabulate_results}}.}

\item{.majority}{A numeric scalar between 0 and 1. The maximum proportion of
first-choice votes among nominees.}

\item{.name}{String. The column name containing nominees' names.}

\item{.pct}{String. The column name containing nominees' first-choice vote
proportions.}
}
\value{
A string. An announcement of the nominee earning the maximum
proportion of first-choice votes.
}
\description{
Find the winner and print the announcement.
}
