% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_funs.R
\name{identify_winner}
\alias{identify_winner}
\title{identify_winner}
\usage{
identify_winner(.data, .name = "name", .pct = "pct")
}
\arguments{
\item{.data}{Dataframe or tibble. Nominees in columns; rankings in rows.}

\item{.name}{String. The column name containing nominees' names.}

\item{.pct}{String. The column name containing nominees' first-choice vote
proportions.}
}
\value{
A list of string elements - the announcement(s) printed by
\code{\link{print_max_result}}.
}
\description{
Iterate over steps 2-4 until a nominee has a 50% + 1 majority of first-choice
votes.
}
