% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_funs.R
\name{import_data}
\alias{import_data}
\title{import_data}
\usage{
import_data(
  path,
  item_text = paste0("Rank the nominees to vote for SMEP President as follows:\\n1 = ",
    "first choice...5 = last choice\\nStart by clicking or tapping on a ",
    "name, and a number indicating its rank will appear to the left. ",
    "Rearrange the names by dragging them in your preferred rank order. - ")
)
}
\arguments{
\item{path}{String. Path to Qualtrics CSV file.}

\item{item_text}{String. Text of Qualtrics item prompt that should be
deleted.}
}
\value{
A tibble containing nominee rankings: nominees in columns; rankings
in rows.
}
\description{
Import CSV file from Qualtrics.
}
