% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomination_funs.R
\name{import_nominations}
\alias{import_nominations}
\title{import_nominations}
\usage{
import_nominations(path, item_text = "QID", col_name = "Nominees")
}
\arguments{
\item{path}{String. Path to Qualtrics CSV file.}

\item{item_text}{String. Text of Qualtrics item prompt.}

\item{col_name}{String. Name to be passed to the column of nominees.}
}
\value{
A tibble containing nominees.
}
\description{
Import CSV file from Qualtrics.
}
