% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_funs.R
\name{redistribute_votes}
\alias{redistribute_votes}
\title{redistribute_votes}
\usage{
redistribute_votes(.data, .remove)
}
\arguments{
\item{.data}{Dataframe or tibble. Nominees in columns; rankings in rows.}

\item{.remove}{A string. The name of the nominee with the lowest number of
first-choice votes.}
}
\value{
A tibble containing updated nominee rankings: nominees in columns;
rankings in rows.
}
\description{
If no simple majority exists, remove the nominee with the least first
votes & redistribute among remaining nominees.
}
