% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_funs.R
\name{import_data}
\alias{import_data}
\title{import_data}
\usage{
import_data(
  path,
 
    item_text = paste0("Rearrange the names by dragging them in your preferred rank order. - ")
)
}
\arguments{
\item{path}{String. Path to Qualtrics CSV file.}

\item{item_text}{String. Text of Qualtrics item prompt that should be
deleted.}
}
\value{
A tibble containing nominee rankings: nominees in columns; rankings
in rows.
}
\description{
Import CSV file from Qualtrics.
}
